Add base pay with card UI.
AddCardView is a separate view conponent that can be used in different xib/storyboard with interface builder.
Base view controller for PAyWithCard and AddCard view controllers.
Toast is part of the the base view controller and new them configuration to put it in the top or bottom.
First proposal for the new api for pay with card.
presenter can now present view with actual size.